+++ src/connections/protocol/data_plane.rs
@@ -269,145 +269,4 @@ impl RangeProgressSignal {
     }
 }
 
-/// A single key-value record produced by the client.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ClientRecord {
-    pub key: Vec<u8>,
-    pub value: Vec<u8>,
-}
 
-impl ClientRecord {
-    pub fn new(key: impl Into<Vec<u8>>, value: impl Into<Vec<u8>>) -> Self {
-        Self {
-            key: key.into(),
-            value: value.into(),
-        }
-    }
-
-    /// Serialize a slice of records into a byte buffer.
-    pub fn serialize_batch(records: &[ClientRecord]) -> Vec<u8> {
-        let mut buf = Vec::new();
-        for r in records {
-            buf.extend_from_slice(&(r.key.len() as u32).to_be_bytes());
-            buf.extend_from_slice(&r.key);
-            buf.extend_from_slice(&(r.value.len() as u32).to_be_bytes());
-            buf.extend_from_slice(&r.value);
-        }
-        buf
-    }
-
-    /// Deserialize a byte buffer into a vector of records.
-    pub fn deserialize_batch(
-        mut buf: &[u8],
-        count: u32,
-    ) -> Result<Box<[ClientRecord]>, std::io::Error> {
-        let mut records = Vec::with_capacity(count as usize);
-        for _ in 0..count {
-            if buf.len() < 4 {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "Unexpected EOF reading key length",
-                ));
-            }
-            let key_len = u32::from_be_bytes(buf[..4].try_into().unwrap()) as usize;
-            buf = &buf[4..];
-
-            if buf.len() < key_len {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "Unexpected EOF reading key data",
-                ));
-            }
-            let key = buf[..key_len].to_vec();
-            buf = &buf[key_len..];
-
-            if buf.len() < 4 {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "Unexpected EOF reading value length",
-                ));
-            }
-            let val_len = u32::from_be_bytes(buf[..4].try_into().unwrap()) as usize;
-            buf = &buf[4..];
-
-            if buf.len() < val_len {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "Unexpected EOF reading value data",
-                ));
-            }
-            let value = buf[..val_len].to_vec();
-            buf = &buf[val_len..];
-
-            records.push(ClientRecord { key, value });
-        }
-        Ok(records.into_boxed_slice())
-    }
-}
-
-/// Compression codec for opaque entry payloads.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, BorshSerialize, BorshDeserialize)]
-#[borsh(use_discriminant = true)]
-#[repr(u8)]
-pub enum CompressionCodec {
-    None = 0,
-    Lz4 = 1,
-    Zstd = 2,
-}
-
-impl CompressionCodec {
-    pub fn from_u8(tag: u8) -> Result<Self, std::io::Error> {
-        match tag {
-            0 => Ok(CompressionCodec::None),
-            1 => Ok(CompressionCodec::Lz4),
-            2 => Ok(CompressionCodec::Zstd),
-            other => Err(std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!("Unknown compression codec tag: {}", other),
-            )),
-        }
-    }
-
-    pub fn compress(&self, data: &[u8]) -> Result<Vec<u8>, std::io::Error> {
-        match self {
-            CompressionCodec::None => Ok(data.to_vec()),
-            CompressionCodec::Lz4 => Ok(lz4_flex::compress_prepend_size(data)),
-            CompressionCodec::Zstd => zstd::encode_all(data, 0),
-        }
-    }
-
-    pub fn decompress(&self, data: &[u8]) -> Result<Vec<u8>, std::io::Error> {
-        match self {
-            CompressionCodec::None => Ok(data.to_vec()),
-            CompressionCodec::Lz4 => lz4_flex::decompress_size_prepended(data)
-                .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e.to_string())),
-            CompressionCodec::Zstd => zstd::decode_all(data),
-        }
-    }
-
-    /// Helper to encode a batch of records into an opaque EntryPayload with a 1-byte codec tag.
-    pub fn encode_payload(&self, records: &[ClientRecord]) -> Result<Vec<u8>, std::io::Error> {
-        let serialized = ClientRecord::serialize_batch(records);
-        let compressed = self.compress(&serialized)?;
-        let mut payload = Vec::with_capacity(1 + compressed.len());
-        payload.push(*self as u8);
-        payload.extend_from_slice(&compressed);
-        Ok(payload)
-    }
-
-    /// Helper to decode an opaque EntryPayload (with 1-byte codec tag) into records.
-    pub fn decode_payload(
-        payload: &[u8],
-        record_count: u32,
-    ) -> Result<Box<[ClientRecord]>, std::io::Error> {
-        if payload.is_empty() {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::UnexpectedEof,
-                "Empty payload, missing codec tag",
-            ));
-        }
-        let codec = Self::from_u8(payload[0])?;
-        let decompressed = codec.decompress(&payload[1..])?;
-        ClientRecord::deserialize_batch(&decompressed, record_count)
-    }
-}
